Cleanup CommunityPreviewCard
